jldi jldi mai update kr rha hu bcoz abhi kam sahi chal rha hai and payment vaali dikkat abhi bhi h

+++ E-Commerce Project/store/models.py
@@ -12,7 +12,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name=models.CharField(max_length=200)
-    price=models.FloatField()
+    price=models.DecimalField(max_digits=7,decimal_places=3)
     image=models.ImageField(null=True,blank=True)
     digital=models.BooleanField(default=False,null=True,blank=True)
     def __str__(self):
